server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.2
cycler==0.11.0
distlib==0.3.6
dnspython==2.4.2
EasyProcess==1.1
email-validator==2.0.0.post2
entrypoint2==1.1
exceptiongroup==1.1.1
fastapi==0.103.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
MouseInfo==0.1.3
mss==9.0.1
networkx==2.8.6
numpy==1.23.2
orjson==3.9.5
outcome==1.2.0
packaging==21.3
pandas==2.0.1
Pillow==9.2.0
platformdirs==2.5.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
rubicon-objc==0.4.6
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.12
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.16.5
watchfiles==0.20.0
webdriver-manager==3.8.6
websockets==11.0.3
Werkzeug==2.2.2
wsproto==1.2.0
